% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{coerce_num_cln}
\alias{coerce_num_cln}
\title{Convert non-numeric variables and coerce to numeric}
\usage{
coerce_num_cln(string)
}
\arguments{
\item{string}{string}
}
\description{
Convert non-numeric variables and coerce to numeric
}
