% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pca_ind_scores}
\alias{pca_ind_scores}
\title{CFP and organizational charactersitics data}
\format{
A data frame with 1195 rows and 110 variables:
}
\source{
Data gathered from published company annual reports
}
\usage{
pca_ind_scores
}
\description{
A dataset containing variables on CFP, CSR, and company characteristics
for 238 African companies spanning a period of 5 years (2014 to 2018)
}
\details{
Note: some data was used from companies in the period before they
were officially listed  on teh stock exchange
}
\keyword{datasets}
