% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reg_data}
\alias{reg_data}
\title{CFP and organizational charactersitics data}
\format{
A data frame with 1195 rows and 110 variables:

\describe{
\item{Country}{Categorization of industry, 9 categories}
\item{Multinational}{company name}
\item{Sector}{Country of listing}
\item{listing_age}{Whether multinational, binary}
\item{incorp_age}{Specific industry, 30 categories}
\item{lag_return_on_assets}{Year of report, 2014 - 2018}
\item{lag_return_on_equity}{Year company listed on stock exchange, 1927 - 2018}
\item{social_committee}{Year company incorporated, 1852 - 2012}
\item{reporting_year}{Net profit reported, different various currencies}
}
}
\source{
Data gathered from published company annual reports
}
\usage{
reg_data
}
\description{
A dataset containing variables on CFP, CSR, and company characteristics
for 238 African companies spanning a period of 5 years (2014 to 2018)
}
\details{
Note: some data was used from companies in the period before they
were officially listed  on teh stock exchange
}
\keyword{datasets}
