% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rem_zero_var_cols}
\alias{rem_zero_var_cols}
\title{Remove columns which are all the same value for numeric variables}
\usage{
rem_zero_var_cols(dat)
}
\arguments{
\item{dat}{dataframe}
}
\description{
Remove columns which are all the same value for numeric variables
}
