% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_non_num}
\alias{remove_non_num}
\title{Remove non-numeric characters from a string}
\usage{
remove_non_num(string)
}
\arguments{
\item{string}{string}
}
\description{
Remove non-numeric characters from a string
}
