% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{select_num_vars}
\alias{select_num_vars}
\title{Select columns of numeric variables}
\usage{
select_num_vars(dat)
}
\arguments{
\item{dat}{dataframe}
}
\description{
Select columns of numeric variables
}
